Rewrote to full strict TS. Api v3 is only one supported
