--- frontend-examples/api/v1/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-    <script src="./index.js" defer></script>
-  </head>
-  <body></body>
-</html>
